tasks/misc-faceless-void/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
faiss-cpu==1.7.4
torch==2.1.1
torchvision==0.16.1
facenet-pytorch==2.5.3
python-jose[cryptography]==3.3.0
pyjwt==2.8.0 